app/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
black==20.8b1
click==7.1.2
coverage==5.3.1
Django==3.1.4
djangorestframework==3.12.2
dj-database-url==0.5.0
flake8==3.8.4
gunicorn==20.0.4
importlib-metadata==3.1.1
iniconfig==1.1.1
isort==5.6.4
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.7
pathspec==0.8.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
pytz==2020.4
regex==2020.11.13
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
zipp==3.4.0
